# examples/onk-simple-api/api/server.gen.go
// Code generated by onek. DO NOT EDIT.
package api

import (
	"context"
	"crypto/rand"
	"encoding/hex"
	"encoding/json"
	"fmt"
	"net/http"
	"regexp"
	"strconv"
	"time"
)

func writeJSONError(w http.ResponseWriter, status int, message string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(map[string]string{"message": message})
}

func parseInt32(s string) (int32, error) { v, err := strconv.ParseInt(s, 10, 32); return int32(v), err }
func parseInt64(s string) (int64, error) { return strconv.ParseInt(s, 10, 64) }
func parseUint32(s string) (uint32, error) {
	v, err := strconv.ParseUint(s, 10, 32)
	return uint32(v), err
}
func parseUint64(s string) (uint64, error) { return strconv.ParseUint(s, 10, 64) }
func parseFloat32(s string) (float32, error) {
	v, err := strconv.ParseFloat(s, 32)
	return float32(v), err
}
func parseFloat64(s string) (float64, error) { return strconv.ParseFloat(s, 64) }
func parseBool(s string) (bool, error) {
	if s == "true" {
		return true, nil
	}
	if s == "false" {
		return false, nil
	}
	return false, fmt.Errorf("must be true or false")
}

func validHeaderFormat(value, format string) bool {
	switch format {
	case "uuid":
		return regexp.MustCompile(`^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$`).MatchString(value)
	case "email":
		return regexp.MustCompile(`^[^@\s]+@[^@\s]+\.[^@\s]+$`).MatchString(value)
	case "uri":
		return regexp.MustCompile(`^[a-zA-Z][a-zA-Z0-9+.-]*://[^\s]+$`).MatchString(value)
	default:
		return true
	}
}

// RequestMetadata identifies the generated route handling a request.
type RequestMetadata struct {
	Service     string
	Method      string
	HTTPMethod  string
	Route       string
	AuthSchemes []string
}

type requestMetadataContextKey struct{}
type requestIDContextKey struct{}

func RequestMetadataFromContext(ctx context.Context) (RequestMetadata, bool) {
	metadata, ok := ctx.Value(requestMetadataContextKey{}).(RequestMetadata)
	return metadata, ok
}

func RequestIDFromContext(ctx context.Context) (string, bool) {
	requestID, ok := ctx.Value(requestIDContextKey{}).(string)
	return requestID, ok && requestID != ""
}

type Middleware func(http.Handler) http.Handler
type RequestIDGenerator func() string
type Authorizer func(context.Context, RequestMetadata, *http.Request) error
type RequestResult struct {
	StatusCode int
	Duration   time.Duration
}
type RequestObserver interface {
	RequestStarted(context.Context, RequestMetadata) context.Context
	RequestFinished(context.Context, RequestMetadata, RequestResult)
}

type ServerOption func(*serverOptions)

type serverOptions struct {
	mux                *http.ServeMux
	middlewares        []Middleware
	requestIDHeader    string
	requestIDGenerator RequestIDGenerator
	authorizer         Authorizer
	observer           RequestObserver
}

// WithMux supports the options-first registration form.
func WithMux(mux *http.ServeMux) ServerOption { return func(o *serverOptions) { o.mux = mux } }

func WithMiddleware(middleware ...Middleware) ServerOption {
	return func(o *serverOptions) { o.middlewares = append(o.middlewares, middleware...) }
}

func WithRequestID(headerName string) ServerOption {
	return WithRequestIDGenerator(headerName, defaultRequestIDGenerator)
}
func WithRequestIDGenerator(headerName string, generate RequestIDGenerator) ServerOption {
	return func(o *serverOptions) {
		if headerName == "" {
			headerName = "X-Request-ID"
		}
		o.requestIDHeader, o.requestIDGenerator = headerName, generate
	}
}

func WithAuthorizer(authorizer Authorizer) ServerOption {
	return func(o *serverOptions) { o.authorizer = authorizer }
}
func WithRequestObserver(observer RequestObserver) ServerOption {
	return func(o *serverOptions) { o.observer = observer }
}

func defaultRequestIDGenerator() string {
	var value [16]byte
	if _, err := rand.Read(value[:]); err != nil {
		return ""
	}
	return hex.EncodeToString(value[:])
}

func (o serverOptions) wrapHandler(handler http.Handler, metadata RequestMetadata) http.Handler {
	for i := len(o.middlewares) - 1; i >= 0; i-- {
		if o.middlewares[i] != nil {
			handler = o.middlewares[i](handler)
		}
	}
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), requestMetadataContextKey{}, metadata)
		if o.requestIDHeader != "" {
			requestID := r.Header.Get(o.requestIDHeader)
			if requestID == "" && o.requestIDGenerator != nil {
				requestID = o.requestIDGenerator()
			}
			if requestID != "" {
				w.Header().Set(o.requestIDHeader, requestID)
				ctx = context.WithValue(ctx, requestIDContextKey{}, requestID)
			}
		}
		if o.authorizer != nil {
			if err := o.authorizer(ctx, metadata, r); err != nil {
				writeJSONError(w, http.StatusUnauthorized, err.Error())
				return
			}
		}
		if o.observer == nil {
			handler.ServeHTTP(w, r.WithContext(ctx))
			return
		}
		started := time.Now()
		ctx = o.observer.RequestStarted(ctx, metadata)
		rw := &statusResponseWriter{ResponseWriter: w, statusCode: http.StatusOK}
		defer func() {
			o.observer.RequestFinished(ctx, metadata, RequestResult{StatusCode: rw.statusCode, Duration: time.Since(started)})
		}()
		handler.ServeHTTP(rw, r.WithContext(ctx))
	})
}

type statusResponseWriter struct {
	http.ResponseWriter
	statusCode int
}

func (w *statusResponseWriter) WriteHeader(statusCode int) {
	w.statusCode = statusCode
	w.ResponseWriter.WriteHeader(statusCode)
}
func (w *statusResponseWriter) Flush() {
	if flusher, ok := w.ResponseWriter.(http.Flusher); ok {
		flusher.Flush()
	}
}

type UserServiceServer interface {
	CreateUser(ctx context.Context, req *CreateUserRequest) (*User, error)
	GetUser(ctx context.Context, req *GetUserRequest) (*User, error)
	Login(ctx context.Context, req *LoginRequest) (*LoginResponse, error)
}

func RegisterUserServiceServer(first any, rest ...any) error {
	var o serverOptions
	var srv UserServiceServer
	if mux, ok := first.(*http.ServeMux); ok {
		o.mux = mux
		if len(rest) == 0 {
			return fmt.Errorf("RegisterUserServiceServer: implementation is required")
		}
		var ok bool
		srv, ok = rest[0].(UserServiceServer)
		if !ok {
			return fmt.Errorf("RegisterUserServiceServer: invalid implementation")
		}
		rest = rest[1:]
	} else {
		var ok bool
		srv, ok = first.(UserServiceServer)
		if !ok {
			return fmt.Errorf("RegisterUserServiceServer: invalid implementation")
		}
	}
	for _, value := range rest {
		opt, ok := value.(ServerOption)
		if !ok {
			return fmt.Errorf("RegisterUserServiceServer: options must be ServerOption values")
		}
		opt(&o)
	}
	if o.mux == nil {
		return fmt.Errorf("RegisterUserServiceServer: mux is required")
	}
	mux := o.mux
	mux.Handle("POST /api/v1/users", o.wrapHandler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		req := new(CreateUserRequest)
		if r.Body != nil {
			if err := json.NewDecoder(r.Body).Decode(req); err != nil && err.Error() != "EOF" {
				writeJSONError(w, http.StatusBadRequest, "invalid request body: "+err.Error())
				return
			}
		}
		{
			value := r.Header.Get("X-API-Key")
			if value == "" {
				writeJSONError(w, http.StatusBadRequest, "missing required header: X-API-Key")
				return
			}
			if value != "" && !validHeaderFormat(value, "uuid") {
				writeJSONError(w, http.StatusBadRequest, "invalid header X-API-Key: expected uuid")
				return
			}
		}
		if v, ok := any(req).(interface{ Validate() error }); ok {
			if err := v.Validate(); err != nil {
				writeJSONError(w, http.StatusBadRequest, err.Error())
				return
			}
		}
		resp, err := srv.CreateUser(r.Context(), req)
		if err != nil {
			writeJSONError(w, http.StatusInternalServerError, err.Error())
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}), RequestMetadata{Service: "UserService", Method: "createUser", HTTPMethod: "POST", Route: "/api/v1/users", AuthSchemes: nil}))
	mux.Handle("POST /api/v1/users/get", o.wrapHandler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		req := new(GetUserRequest)
		if r.Body != nil {
			if err := json.NewDecoder(r.Body).Decode(req); err != nil && err.Error() != "EOF" {
				writeJSONError(w, http.StatusBadRequest, "invalid request body: "+err.Error())
				return
			}
		}
		{
			value := r.Header.Get("X-API-Key")
			if value == "" {
				writeJSONError(w, http.StatusBadRequest, "missing required header: X-API-Key")
				return
			}
			if value != "" && !validHeaderFormat(value, "uuid") {
				writeJSONError(w, http.StatusBadRequest, "invalid header X-API-Key: expected uuid")
				return
			}
		}
		if v, ok := any(req).(interface{ Validate() error }); ok {
			if err := v.Validate(); err != nil {
				writeJSONError(w, http.StatusBadRequest, err.Error())
				return
			}
		}
		resp, err := srv.GetUser(r.Context(), req)
		if err != nil {
			writeJSONError(w, http.StatusInternalServerError, err.Error())
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}), RequestMetadata{Service: "UserService", Method: "getUser", HTTPMethod: "POST", Route: "/api/v1/users/get", AuthSchemes: nil}))
	mux.Handle("POST /api/v1/auth/login", o.wrapHandler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		req := new(LoginRequest)
		if r.Body != nil {
			if err := json.NewDecoder(r.Body).Decode(req); err != nil && err.Error() != "EOF" {
				writeJSONError(w, http.StatusBadRequest, "invalid request body: "+err.Error())
				return
			}
		}
		{
			value := r.Header.Get("X-API-Key")
			if value == "" {
				writeJSONError(w, http.StatusBadRequest, "missing required header: X-API-Key")
				return
			}
			if value != "" && !validHeaderFormat(value, "uuid") {
				writeJSONError(w, http.StatusBadRequest, "invalid header X-API-Key: expected uuid")
				return
			}
		}
		{
			value := r.Header.Get("X-Request-Id")
			if value == "" {
				writeJSONError(w, http.StatusBadRequest, "missing required header: X-Request-Id")
				return
			}
			if value != "" && !validHeaderFormat(value, "uuid") {
				writeJSONError(w, http.StatusBadRequest, "invalid header X-Request-Id: expected uuid")
				return
			}
		}
		if v, ok := any(req).(interface{ Validate() error }); ok {
			if err := v.Validate(); err != nil {
				writeJSONError(w, http.StatusBadRequest, err.Error())
				return
			}
		}
		resp, err := srv.Login(r.Context(), req)
		if err != nil {
			writeJSONError(w, http.StatusInternalServerError, err.Error())
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}), RequestMetadata{Service: "UserService", Method: "login", HTTPMethod: "POST", Route: "/api/v1/auth/login", AuthSchemes: nil}))
	return nil
}