# examples/onk-simple-api/docs/openapi.yaml
openapi: 3.1.0
info:
    title: OnkSimpleAPI
    description: Example service migrated from examples/simple-api to .onk
    version: 1.0.0
paths:
    /api/v1/users:
        post:
            tags:
                - UserService
            summary: createUser
            description: Create a new user.
            operationId: UserService_createUser
            parameters:
                - name: X-API-Key
                  in: header
                  required: true
                  schema:
                    type: string
                    format: uuid
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/examples.onksimpleapi.models.CreateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/examples.onksimpleapi.models.User'
    /api/v1/users/get:
        post:
            tags:
                - UserService
            summary: getUser
            description: Get a user by ID.
            operationId: UserService_getUser
            parameters:
                - name: X-API-Key
                  in: header
                  required: true
                  schema:
                    type: string
                    format: uuid
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/examples.onksimpleapi.models.GetUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/examples.onksimpleapi.models.User'
    /api/v1/auth/login:
        post:
            tags:
                - UserService
            summary: login
            description: Login with different authentication methods.
            operationId: UserService_login
            parameters:
                - name: X-API-Key
                  in: header
                  required: true
                  schema:
                    type: string
                    format: uuid
                - name: X-Request-Id
                  in: header
                  required: true
                  schema:
                    type: string
                    format: uuid
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/examples.onksimpleapi.models.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/examples.onksimpleapi.models.LoginResponse'
components:
    schemas:
        examples.onksimpleapi.models.User:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                    maxLength: 100
                    minLength: 1
                email:
                    type: string
                    format: email
                created_at:
                    type: string
                    format: int64
            description: A system user.
        examples.onksimpleapi.models.CreateUserRequest:
            type: object
            properties:
                name:
                    type: string
                    maxLength: 100
                    minLength: 2
                email:
                    type: string
                    format: email
        examples.onksimpleapi.models.GetUserRequest:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
        examples.onksimpleapi.models.EmailAuth:
            type: object
            properties:
                email:
                    type: string
                    format: email
                password:
                    type: string
                    maxLength: 72
                    minLength: 8
                use_two_factor:
                    type: boolean
                two_factor_code:
                    type: string
        examples.onksimpleapi.models.TokenAuth:
            type: object
            properties:
                token:
                    type: string
                    maxLength: 256
                    minLength: 10
                token_type:
                    type: string
                issued_at:
                    type: string
                    format: int64
        examples.onksimpleapi.models.SocialAuth:
            type: object
            properties:
                provider:
                    type: string
                    enum:
                        - google
                        - github
                        - facebook
                        - apple
                access_token:
                    type: string
                    maxLength: 512
                    minLength: 20
                redirect_uri:
                    type: string
                state:
                    type: string
                permissions:
                    type: array
                    items:
                        type: string
        examples.onksimpleapi.models.LoginRequest:
            type: object
            properties:
                device_id:
                    type: string
                ip_address:
                    type: string
                retry_count:
                    type: integer
                    format: int32
                user_agent:
                    type: string
                auth_method:
                    oneOf:
                        - $ref: '#/components/schemas/examples.onksimpleapi.models.EmailAuth'
                        - $ref: '#/components/schemas/examples.onksimpleapi.models.TokenAuth'
                        - $ref: '#/components/schemas/examples.onksimpleapi.models.SocialAuth'
                    discriminator:
                        propertyName: auth_type
                        mapping:
                            email: '#/components/schemas/examples.onksimpleapi.models.EmailAuth'
                            token: '#/components/schemas/examples.onksimpleapi.models.TokenAuth'
                            social: '#/components/schemas/examples.onksimpleapi.models.SocialAuth'
                remember_me:
                    type: boolean
                session_id:
                    type: string
                timestamp:
                    type: string
                    format: int64
                scopes:
                    type: array
                    items:
                        type: string
        examples.onksimpleapi.models.LoginResponse:
            type: object
            properties:
                access_token:
                    type: string
                refresh_token:
                    type: string
                expires_in:
                    type: string
                    format: int64
                user:
                    allOf:
                        - $ref: '#/components/schemas/examples.onksimpleapi.models.User'